refactor extraction code into extraction service
